refactored hub into apps/ directory, phase 1
