--- pkg/controller/controller.go
@@ -1,178 +1,0 @@
-/*
-Copyright (C) 2018 Synopsys, Inc.
-
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements. See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership. The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License. You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied. See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-
-package controller
-
-import (
-	"fmt"
-	"time"
-
-	hubclientset "github.com/blackducksoftware/perceptor-protoform/pkg/client/clientset/versioned"
-	log "github.com/sirupsen/logrus"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/apimachinery/pkg/util/wait"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/cache"
-	"k8s.io/client-go/util/workqueue"
-)
-
-// Controller will store the controller configuration
-type Controller struct {
-	logger       *log.Entry
-	clientset    kubernetes.Interface
-	queue        workqueue.RateLimitingInterface
-	informer     cache.SharedIndexInformer
-	handler      Handler
-	hubclientset *hubclientset.Clientset
-	namespace    string
-}
-
-// Run will be executed to create the informers or controllers
-func (c *Controller) Run(threadiness int, stopCh <-chan struct{}) {
-	// handle a panic with logging and exiting
-	defer runtime.HandleCrash()
-
-	// ignore new items in the queue but when all goroutines
-	// have completed existing items then shutdown
-	defer c.queue.ShutDown()
-
-	c.logger.Info("Initiating controller")
-
-	// run the informer to start listing and watching resources
-	go c.informer.Run(stopCh)
-
-	// do the initial synchronization (one time) to populate resources
-	if !cache.WaitForCacheSync(stopCh, c.HasSynced) {
-		runtime.HandleError(fmt.Errorf("Error syncing cache"))
-		return
-	}
-	c.logger.Info("Controller cache sync complete")
-
-	for i := 0; i < threadiness; i++ {
-		// run the runWorker method every second with a stop channel
-		go wait.Until(c.runWorker, time.Second, stopCh)
-	}
-
-	// Wait until we're told to stop
-	<-stopCh
-}
-
-// HasSynced will check for informer sync
-func (c *Controller) HasSynced() bool {
-	return c.informer.HasSynced()
-}
-
-// runWorker executes the loop to process new items added to the queue
-func (c *Controller) runWorker() {
-	log.Debug("Controller.runWorker: starting")
-
-	// invoke processNextItem to fetch and consume the next change
-	// to a watched or listed resource
-	for c.processNextItem() {
-		log.Info("Controller.runWorker: processing next item")
-	}
-
-	log.Debug("Controller.runWorker: completed")
-}
-
-// processNextItem retrieves each queued item and takes the
-// necessary handler action based off of if the item was
-// created or deleted
-func (c *Controller) processNextItem() bool {
-	log.Info("Controller.processNextItem: start")
-
-	// fetch the next item (blocking) from the queue to process or
-	// if a shutdown is requested then return out of this to stop
-	// processing
-	key, quit := c.queue.Get()
-
-	// stop the worker loop from running as this indicates we
-	// have sent a shutdown message that the queue has indicated
-	// from the Get method
-	if quit {
-		return false
-	}
-
-	defer c.queue.Done(key)
-
-	// assert the string out of the key (format `namespace/name`)
-	keyRaw := key.(string)
-
-	// take the string key and get the object out of the indexer
-	//
-	// item will contain the complex object for the resource and
-	// exists is a bool that'll indicate whether or not the
-	// resource was created (true) or deleted (false)
-	//
-	// if there is an error in getting the key from the index
-	// then we want to retry this particular queue key a certain
-	// number of times (5 here) before we forget the queue key
-	// and throw an error
-	item, exists, err := c.informer.GetIndexer().GetByKey(keyRaw)
-	if err != nil {
-		if c.queue.NumRequeues(key) < 5 {
-			c.logger.Errorf("Controller.processNextItem: Failed processing item with key %s with error %v, retrying", key, err)
-			c.queue.AddRateLimited(key)
-		} else {
-			c.logger.Errorf("Controller.processNextItem: Failed processing item with key %s with error %v, no more retries", key, err)
-			c.queue.Forget(key)
-			runtime.HandleError(err)
-		}
-	}
-
-	// if the item doesn't exist then it was deleted and we need to fire off the handler's
-	// ObjectDeleted method. but if the object does exist that indicates that the object
-	// was created (or updated) so run the ObjectCreated method
-	//
-	// after both instances, we want to forget the key from the queue, as this indicates
-	// a code path of successful queue key processing
-	if !exists {
-		c.logger.Debugf("Controller.processNextItem: object deleted detected: %s: %+v", keyRaw, item)
-		c.handler.ObjectDeleted(keyRaw)
-		c.queue.Forget(key)
-	} else {
-		c.logger.Debugf("Controller.processNextItem: object created detected: %s: %+v", keyRaw, item)
-		c.handler.ObjectCreated(item)
-		c.queue.Forget(key)
-	}
-
-	// keep the worker loop running by returning true
-	return true
-}
-
-// HubNamespaces will list the hub namespaces
-func (c *Controller) HubNamespaces() ([]string, error) {
-	// 1. get Hub CDR list from default ns
-	hubList, err := c.hubclientset.SynopsysV1().Hubs(c.namespace).List(metav1.ListOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	// 2. extract the namespaces
-	hubNamespaces := []string{}
-	for _, hub := range hubList.Items {
-		if len(hub.Spec.Namespace) > 0 {
-			hubNamespaces = append(hubNamespaces, hub.Spec.Namespace)
-		}
-	}
-	return hubNamespaces, nil
-}
